+++ include/rain.hpp
@@ -4,7 +4,7 @@
 // Versioning.
 #define RAIN_VERSION_MAJOR 7
 #define RAIN_VERSION_MINOR 2
-#define RAIN_VERSION_REVISION 32
+#define RAIN_VERSION_REVISION 33
 
 // Disable secure warnings; the caller should be aware of out-of-bounds errors
 // when calling any not-bound-checked function.
@@ -17,7 +17,6 @@
 #include "rain/algorithm.hpp"
 #include "rain/error.hpp"
 #include "rain/filesystem.hpp"
-#include "rain/gdi-plus.hpp"
 #include "rain/literal.hpp"
 #include "rain/log.hpp"
 #include "rain/multithreading.hpp"