+++ include/rain/windows.hpp
@@ -2,4 +2,6 @@
 #pragma once
 
 #include "windows/exception.hpp"
+#include "windows/gdiplus.hpp"
+#include "windows/theme.hpp"
 #include "windows/windows.hpp"