+++ include/rain/windows/gdiplus.hpp
@@ -1,0 +1,4 @@
+#pragma once
+
+#include "gdiplus/exception.hpp"
+#include "gdiplus/gdiplus.hpp"